Call low-level functions given data

This is the really "entry point" function that gives a convenient access to all the low-level pre-processing functions. These low-level functions are also exported and documented with the details of the returned values.

preproc_data(data, prep_fun_name, by = NULL, ..., character.only = FALSE)

Arguments

data
Raw data of class data.frame.

prep_fun_name
The name of a function, given as a name or literal character string, depending on whether character.only is FALSE (default) or TRUE. Only functions from this package is supported for now.

by
The column name(s) in data used to be grouped by. If set to NULL (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.

...
These dots are for future extensions and must be empty.

character.only
A logical indicating whether prep_fun_name can be assumed to be character strings.

Value

A tibble of game performances returned by low-level functions.

Author

Liang Zhang psychelzh@outlook.com