Task Switching and Stroop-like paradigm

In task switching paradigms, two types of tasks switch between each other, so the "switch cost" can be calculated (using switchcost()). Similarly, in Stroop-like tasks, stimuli are classified into two conditions (i.e., "congruent" and "incongruent"), so the "congruence effect" can be calculated (using congeff()). There are also special types of tests where congruence effect and switch cost both exist, from which complexswitch() calculates both.

complexswitch(data, by, vars_input)

congeff(data, by, vars_input)

switchcost(data, by, vars_input)

Arguments

data
Raw data of class data.frame.

by
The column name(s) in data used to be grouped by.

vars_input
The matched variable names for further processing. Typically set by the high-level function preproc_data().

Value

A tibble with the following variables:

For congruence effect:

mrt_con

Mean reaction time for congruent trials.

mrt_inc

Mean reaction time for incogruent trials.

cong_eff_rt

Congruence effect of reaction time (RT), i.e., incongruent RT - congruent RT.

pc_con

Percent of correct for congruent trials.

pc_inc

Percent of correct for incogruent trials.

cong_eff_pc

Congruency effect of percent of correct (PC), i.e., congruent PC - incongruent PC.

For switch cost:
mrt_pure

Mean reaction time for non-mixed blocks.

mrt_repeat

Mean reaction time for repeat trials.

mrt_switch

Mean reaction time for switch trials.

switch_cost_rt_gen

General switch cost (based on mean reaction times).

switch_cost_rt_spe

Specific switch cost (based on mean reaction times).

pc_pure

Percent of correct for non-mixed blocks.

pc_repeat

Percent of correct for repeat trials.

pc_switch

Percent of correct for switch trials.

switch_cost_pc_gen

General switch cost (based on percent of correct).

switch_cost_pc_spe

Specific switch cost (based on percent of correct).